pexels/requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
urllib3>=2.0.0
